added middleware for authentication
