+++ api/controllers/users.js
@@ -1,0 +1,104 @@
+const mongoose = require('mongoose');
+const bcrypt = require('bcrypt');
+const jwt = require('jsonwebtoken');
+const User = require('../models/users');
+
+exports.signup = (req,res,next) =>{
+    User.find({email: req.body.email}).exec()
+    .then(user => {
+        if(user.length >= 1) {
+            return res.status(420).json({
+                message:'email already exists'
+            });
+        }else{
+            bcrypt.hash(req.body.password,10,(err,hash) => {
+                if(err){
+                    return res.status(500).json({
+                        error: err
+                    });
+                }else{
+                    const user = new User({
+                        _id: mongoose.Types.ObjectId(),
+                        email: req.body.email,
+                        password: hash
+                    });
+                    user
+                    .save()
+                    .then(result => {
+                        console.log(result);
+                        res.status(201).json({
+                            message: 'User added'
+                        });
+                    })
+                    .catch(err => {
+                        console.log(err);
+                        res.status(500).json({
+                            error: err
+                        });
+                    });
+                }
+            });
+        }
+    });
+}
+
+
+exports.login = (req,res,next) => {
+    User.find({email: req.body.email})
+    .exec()
+    .then(user => {
+        if(user.length < 1){
+            return res.status(404).json({
+                message: 'Auth failed'
+            });
+        }
+        bcrypt.compare(req.body.password, user[0].password,(err,result) => {
+            if(err){
+                return res.status(404).json({
+                    message: 'Auth failed'
+                });
+            }
+            if(result){
+                const token = jwt.sign(
+                    {email: user[0].email, 
+                     userId: user[0]._id
+                    },
+                    "Vaibhav",
+                    {
+                        expiresIn: "1h"
+                    }
+                );
+                return res.status(200).json({
+                    message: 'Auth successful',
+                    token: token
+                });
+            }
+            return res.status(404).json({
+                message: 'Auth failed'
+            });
+        });
+    })
+    .catch(err => {
+        console.log(err);
+        res.status(500).json({
+            error: err
+        });
+    });
+}
+
+
+exports.deleteUser = (req,res,next) => {
+    User.remove({_id: req.params.userId})
+    .exec()
+    .then(result => {
+        res.status(200).json({
+            message:'User deleted successfully'
+        });
+    })
+    .catch(err => {
+        console.log(err);
+        res.status(500).json({
+            error: err
+        });
+    });
+}
