+++ api/models/users.js
@@ -1,0 +1,16 @@
+const mongoose = require('mongoose');
+
+const userSchema = mongoose.Schema({
+    _id: mongoose.Schema.Types.ObjectId,
+    email: {
+        type: String,
+        required: true,
+        unique:true,
+    },
+    password: {
+        type: String, 
+        required: true
+    }
+})
+
+module.exports = mongoose.model('User', userSchema);
