+++ api/routes/orders.js
@@ -1,121 +1,16 @@
 const express = require('express');
 const router = express.Router();
-const mongoose = require('mongoose');
+const checkAuth = require('../middleware/check-auth');
 
-const Order = require('../models/order');
-const Product = require('../models/product');
+const OrdersController = require('../controllers/orders');
 
-router.get('/', (req, res, next) => {
-    Order
-        .find()
-        .select('product quantity _id')
-        .populate('product')
-        .exec()
-        .then(docs => {
-            res.status(200).json({
-                count: docs.length,
-                orders: docs.map(doc => {
-                    return{
-                        _id: doc._id,
-                        product: doc.productId,
-                        quantity: doc.quantity,
-                        request:{
-                            type: 'GET',
-                            url: 'http://orders/'+doc._id
-                        }
-                    }
-                })
-            });
-        })
-        .catch(err => {
-            res.status(500).json({
-                error: err
-            });
-        })
-});
+router.get('/',checkAuth,OrdersController.allOrder);
 
-router.post('/', (req, res, next) => {
-    
-    Product.findById(req.body.productId).then(product => {
-        if(!product){
-            return res.status(404).json({
-                message: 'Product not found'
-            });
-        }
-    const order = new Order({
-            _id: mongoose.Types.ObjectId(),
-            quantity: req.body.quantity,
-            product: req.body.productId
-    });
-    return order
-      .save()
-      .then(result =>{
-          console.log(result);
-          res.status(201).json({
-              message: 'Order stored',
-              createdOrder:{
-                  _id: result._id,
-                  product: result.product,
-                  quantity: result.quantity
-              },
-              request:{
-                  type: 'GET',
-                  url: 'http://localhost:3000/orders/'+result._id
-              }
-          });
-      })
-    .catch(err => {
-        res.status(500).json({
-            message: 'Product not found',
-            error: err
-        })
-    });
-  });
-});
+router.post('/',checkAuth,OrdersController.createOrder);
 
 
-router.get('/:orderId', (req, res, next) => {
-    Order.findById(req.params.orderId)
-        .populate('product')
-        .exec()
-        .then(order => {
-            if(!order){
-                return res.status(404).json({
-                    message: 'Order not found'
-                });
-            }
-            res.status(200).json({
-                order: order,
-                request:{
-                    type: 'GET',
-                    url: 'http://localhost:3000/orders/'
-                }
-            });
-        })
-        .catch(err => {
-            res.status(500).json({
-                error: err
-            });
-        });
-});
+router.get('/:orderId',checkAuth, OrdersController.getOrder);
 
-router.delete('/:orderId', (req, res, next) => {
-    Order.remove({_id: req.params.orderId}).exec()
-    .then(result => {
-        res.status(200).json({
-            message:'Order deleted successfully',
-            request:{
-                type:"POST",
-                url:"http://localhost:3000/orders",
-                body:{productId: 'ID', quantity: 'Number'}
-            }
-        })
-    })        
-    .catch(err => {
-        res.status(500).json({
-            error: err
-        });
-    });
-});
+router.delete('/:orderId',checkAuth, OrdersController.deleteOrder);
 
 module.exports = router;
