+++ api/routes/users.js
@@ -1,16 +1,12 @@
 const express = require('express');
 const router = express.Router();
 
-router.get('/', (req,res,next) =>{
-    res.status(200).json({
-        message: 'Handling GET request to /users'
-    });
-});
+const UsersController = require('../controllers/users');
 
-router.post('/', (req,res,next) =>{
-    res.status(201).json({
-        message: 'Handling POST request to /users'
-    });
-});
+router.post('/signup', UsersController.signup);
+
+router.post('/login',UsersController.login);
+
+router.delete('/:userId',UsersController.deleteUser);
 
 module.exports = router;
