+++ package.json
@@ -17,8 +17,10 @@
   },
   "homepage": "https://github.com/Vaibhav-afk/API#readme",
   "dependencies": {
+    "bcrypt": "^5.0.1",
     "body-parser": "^1.19.0",
     "express": "^4.17.1",
+    "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.12.9",
     "morgan": "^1.10.0",
     "nodemon": "^2.0.7"
